backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchsummary==1.5.1
torchvision==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
